refact: community home 화면 훅 분리
