+++ src/App.js
@@ -17,7 +17,7 @@ import AccountManage from './pages/mypage/account/AccountManage';
 import CommunityNoti from './pages/mypage/notification/CommunityNoti';
 import HomeFeed from './pages/home/HomeFeed';
 import HomeItem from './pages/homeItem/HomeItem';
-import Community from './pages/community/main/Community';
+import Community from './pages/community/home/Community';
 import CommunityInfo from './pages/community/detail/CommunityInfo';
 import Category from './pages/ranking/category/Category';
 import Empty from './pages/empty/Empty';
